Factor out state entity fixture loader commonalities
