+++ src/Services/FixtureLoader/TestStateFixtureLoader.php
@@ -3,74 +3,16 @@
 namespace App\Services\FixtureLoader;
 
 use App\Entity\TestState;
-use App\Services\DataProviderInterface;
-use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Component\Console\Output\OutputInterface;
 
-class TestStateFixtureLoader extends AbstractFixtureLoader implements FixtureLoaderInterface
+class TestStateFixtureLoader extends AbstractStateFixtureLoader implements FixtureLoaderInterface
 {
-    public function __construct(EntityManagerInterface $entityManager, DataProviderInterface $dataProvider)
-    {
-        parent::__construct($entityManager, $dataProvider);
-    }
-
     protected function getEntityClass(): string
     {
         return TestState::class;
     }
 
-    public function load(?OutputInterface $output = null): void
+    protected function createEntity(string $name): object
     {
-        if ($output) {
-            $output->writeln('Importing test states ...');
-        }
-
-        foreach ($this->data as $name) {
-            if ($output) {
-                $output->write(sprintf(
-                    '  ' . '<comment>%s</comment> %s...',
-                    $name,
-                    $this->createPostNamePadding($this->data, $name)
-                ));
-            }
-
-            $entity = $this->repository->findOneBy([
-                'name' => $name,
-            ]);
-
-            if (!$entity) {
-                if ($output) {
-                    $output->write(' <fg=cyan>creating</>');
-                }
-
-                $entity = TestState::create($name);
-                $this->entityManager->persist($entity);
-                $this->entityManager->flush();
-            }
-
-            if ($output) {
-                $output->writeln(' <info>✓</info>');
-            }
-        }
-
-        if ($output) {
-            $output->writeln('');
-        }
-    }
-
-    /**
-     * @param string[] $names
-     * @param string $name
-     *
-     * @return string
-     */
-    private function createPostNamePadding(array $names, string $name): string
-    {
-        $lengths = array_map(fn (string $name) => strlen($name), $names);
-
-        $max = max($lengths);
-        $difference = $max - strlen($name);
-
-        return str_repeat(' ', $difference);
+        return TestState::create($name);
     }
 }
